requirements.txt
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
